# docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Fund Management System",
    "description" : "The money surface of a broking account.\n\nTwo conventions hold across every endpoint:\n\n**Money is an integer count of paise, never a decimal.** Rule R5 of\nthe ratified event taxonomy and HLD §9.1c. A float in a published\nschema propagates into every generated client, and those are not\nours to fix.\n\n**Errors are branched on by `code`, never by `message`.** The code\nis stable and safe to send to a client; the message is a\ndeveloper-facing explanation and is not user copy.",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "platformAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/api/v1/funds/transactions" : {
      "get" : {
        "tags" : [ "Transactions" ],
        "summary" : "Transactions for a period, in one view",
        "description" : "Rule L5 separates \"where is my money\" from \"explain my account\".\n`MOVEMENTS` carries only money the trader moved in or out; `ALL_ENTRIES` carries\nevery entry with its running balance. Rule L5a is the line that is easy to get\nwrong: sale proceeds and charges are not payins.\n\nThe period defaults to the last 30 days (Rule L6) — chosen because the mandated\nreturn of unused funds runs monthly or quarterly and is among the most-queried\nentries, so a shorter default shows an empty table for a transaction the trader\nknows happened.\n\nAn empty period returns an explicit empty result with the period echoed and a\nwider one suggested (Rule L7), never a bare empty array.",
        "operationId" : "list",
        "parameters" : [ {
          "name" : "view",
          "in" : "query",
          "description" : "MOVEMENTS or ALL_ENTRIES. Defaults to MOVEMENTS, the common question.",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "MOVEMENTS",
            "enum" : [ "MOVEMENTS", "ALL_ENTRIES" ]
          }
        }, {
          "name" : "from",
          "in" : "query",
          "description" : "Inclusive. Omit both bounds for Rule L6's 30-day default.",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Entries, possibly none",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TransactionsResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "`invalid_request` — an inverted period, or one wider than the back office's ledger can answer in a single call",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503" : {
            "description" : "`upstream_unavailable`",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/funds/transactions/{reference}" : {
      "get" : {
        "tags" : [ "Transactions" ],
        "summary" : "One entry in full",
        "description" : "Looked up across ALL_ENTRIES regardless of view, so an entry the movements view\nfilters out is still reachable by reference.\n\nAn entry belonging to another trader answers 404 rather than 403 — confirming\nexistence would itself leak.",
        "operationId" : "detail",
        "parameters" : [ {
          "name" : "reference",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "from",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The entry",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EntryDto"
                }
              }
            }
          },
          "400" : {
            "description" : "`invalid_request` — a period bound that will not parse as a date. The refusal names the parameter and does not echo the submitted value",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "No such entry for this account, in this period",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/funds/statement.csv" : {
      "get" : {
        "tags" : [ "Transactions" ],
        "summary" : "The same view and period, as CSV (REQ-407)",
        "description" : "Rule L8a: an export returns precisely what is on screen — the same view, the\nsame period, the same running balance.\n\nAmounts are plain unformatted decimals with no currency symbol and no thousands\nseparator, so the file is summable in a spreadsheet without cleaning. The type\ncolumn reads Debit or Credit rather than an internal kind, because the file is\nread against a bank statement.\n\nNo unmasked account number appears anywhere (Profile PR-32); a field that could\ncontain one fails the export rather than being redacted.",
        "operationId" : "statement",
        "parameters" : [ {
          "name" : "view",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "ALL_ENTRIES",
            "enum" : [ "MOVEMENTS", "ALL_ENTRIES" ]
          }
        }, {
          "name" : "from",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "The statement",
            "content" : {
              "text/csv" : { }
            }
          },
          "400" : {
            "description" : "`invalid_request`",
            "content" : {
              "text/csv" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "MoneyDto" : {
        "type" : "object",
        "description" : "A monetary amount as an integer number of paise. Never a decimal or a float.",
        "properties" : {
          "paise" : {
            "type" : "integer",
            "format" : "int64",
            "description" : "Amount in paise. 12345 is ₹123.45.",
            "example" : 12345
          },
          "currency" : {
            "type" : "string",
            "description" : "ISO 4217 currency code.",
            "example" : "INR"
          }
        }
      },
...
      "ErrorResponse" : {
        "type" : "object",
        "description" : "The uniform error body. Clients branch on `code`, never on `message`.",
        "properties" : {
          "code" : {
            "type" : "string",
            "description" : "Stable machine-readable code.",
            "example" : "amount_exceeds_withdrawable"
          },
          "message" : {
            "type" : "string",
            "description" : "Developer-facing explanation. Not user copy."
          },
          "details" : {
            "type" : "object",
            "additionalProperties" : { },
            "description" : "Values needed to explain the refusal without re-fetching."
          }
        }
      },
...
      "EntryDto" : {
        "type" : "object",
        "description" : "One entry as the trader sees it.",
        "properties" : {
          "reference" : {
            "type" : "string",
            "example" : "VCH-4471"
          },
          "date" : {
            "type" : "string",
            "format" : "date"
          },
          "kind" : {
            "type" : "string",
            "example" : "PAYIN"
          },
          "descriptionKey" : {
            "type" : "string",
            "example" : "ENTRY_PAYIN"
          },
          "descriptionParameters" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "secondaryDetail" : {
            "type" : "string"
          },
          "amount" : {
            "$ref" : "#/components/schemas/MoneyDto"
          },
          "direction" : {
            "type" : "string",
            "description" : "IN or OUT"
          },
          "runningBalance" : {
            "$ref" : "#/components/schemas/MoneyDto"
          },
          "segment" : {
            "type" : "string",
            "example" : "NSE_CASH"
          },
          "userCaused" : {
            "type" : "boolean"
          },
          "reversedBy" : {
            "type" : "string"
          },
          "reverses" : {
            "type" : "string"
          }
        }
      },
      "PeriodDto" : {
        "type" : "object",
        "description" : "An inclusive date range.",
        "properties" : {
          "from" : {
            "type" : "string",
            "format" : "date"
          },
          "to" : {
            "type" : "string",
            "format" : "date"
          }
        }
      },
      "TransactionsResponse" : {
        "type" : "object",
        "description" : "Transactions for a period, in one of Rule L5's two views.",
        "properties" : {
          "view" : {
            "type" : "string",
            "description" : "MOVEMENTS or ALL_ENTRIES",
            "example" : "MOVEMENTS"
          },
          "period" : {
            "$ref" : "#/components/schemas/PeriodDto",
            "description" : "The period actually covered, echoed back."
          },
          "entries" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/EntryDto"
            }
          },
          "suggestedWiderPeriod" : {
            "$ref" : "#/components/schemas/PeriodDto",
            "description" : "A wider period worth offering. Present only when entries is empty."
          }
        }
      },
...
    },
    "securitySchemes" : {
      "platformAuth" : {
        "type" : "http",
        "description" : "HTTP Basic, which is what the service enforces today.\n\n**Provisional.** Security review MEDIUM-2 is open on this: Basic carries credentials on every request and has no expiry, revocation, rotation or scope. It is here because a filter chain that refuses beats a correct scheme that does not exist, and it is declared here because a specification naming a scheme the service does not accept produces clients that 401 on every call.\n\nThe deployment is expected to place a gateway in front that validates a platform-issued token. When that issuer is known this scheme becomes `bearer`; the requirement name does not change.",
        "scheme" : "basic"
      }
    }
  }
}